requirements.txt
openai
pandas
python-dotenv
requests
spotipy